It seems CUDA has not overloaded 'abs(short)' device function. I wonder how could it be compiled. Could it be that cluster's gcc has no overloaded abs(short) either and so CUDA can't pick it and uses abs(int) for device then?.

+++ MS4/kernel/gpu/scatter_grid/scatter_grid.cu
@@ -139,7 +139,7 @@ void scatter_grid_point
     }
 
     // Load GCF pixel
-    short supp = abs(uvo[i].gcf_supp);
+    short supp = short(abs(int(uvo[i].gcf_supp))); // CUDA has no overloadeed abs(short) for device.
     complexd px = uvo[i].gcf[myConvU * supp + myConvV];
     px.y = copysign(px.y, double(uvo[i].gcf_supp));
